requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==23.3.0
celery[redis]==5.2.7
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
distlib==0.3.6
Django==4.2
django-extensions==3.2.1
django-filter==23.1
django-guardian==2.4.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
filelock==3.11.0
hyperlink==21.0.0
identify==2.5.22
idna==3.4
incremental==22.10.0
kombu==5.2.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
zope.interface==6.0
